@@ -1,12 +1,13 @@
 # This migration comes from blacklight (originally 20140202020202)
-# -*- encoding : utf-8 -*-
+# frozen_string_literal: true
 class CreateBookmarks < ActiveRecord::Migration
   def self.up
     create_table :bookmarks do |t|
-      t.integer :user_id, :null=>false
+      t.integer :user_id, index: true, null: false
       t.string :user_type
-      t.string :document_id
-      t.string :title
+      t.string :document_id, index: true
+      t.string :document_type
+      t.binary :title
       t.timestamps null: false
     end
   end
