@@ -1,6 +1,7 @@
 class ApplicationController < ActionController::Base
-  # Prevent CSRF attacks by raising an exception.
-  # For APIs, you may want to use :null_session instead.
+  # Adds a few additional behaviors into the application controller
+  include Blacklight::Controller
+  layout 'blacklight'
+
   protect_from_forgery with: :exception
-  config.encoding = "utf-8"
 end
