@@ -1,21 +1,34 @@
-# -*- encoding : utf-8 -*-
-class SolrDocument 
+# frozen_string_literal: true
+class SolrDocument
+  include Blacklight::Solr::Document    
+      # The following shows how to setup this blacklight document to display marc documents
+  extension_parameters[:marc_source_field] = :marc_display
+  extension_parameters[:marc_format_type] = :marcxml
+  use_extension( Blacklight::Solr::Document::Marc) do |document|
+    document.key?( :marc_display  )
+  end
+  
+  field_semantics.merge!(    
+                         :title => "title_display",
+                         :author => "author_display",
+                         :language => "language_facet",
+                         :format => "format"
+                         )
 
-  include Blacklight::Solr::Document
+
 
   # self.unique_key = 'id'
-  
+
   # Email uses the semantic field mappings below to generate the body of an email.
-  SolrDocument.use_extension( Blacklight::Document::Email )
-  
+  SolrDocument.use_extension(Blacklight::Document::Email)
+
   # SMS uses the semantic field mappings below to generate the body of an SMS email.
-  SolrDocument.use_extension( Blacklight::Document::Sms )
+  SolrDocument.use_extension(Blacklight::Document::Sms)
 
   # DublinCore uses the semantic field mappings below to assemble an OAI-compliant Dublin Core document
   # Semantic mappings of solr stored fields. Fields may be multi or
   # single valued. See Blacklight::Document::SemanticFields#field_semantics
   # and Blacklight::Document::SemanticFields#to_semantic_values
   # Recommendation: Use field names from Dublin Core
-  use_extension( Blacklight::Document::DublinCore)    
-
+  use_extension(Blacklight::Document::DublinCore)
 end
