@@ -1,3 +1,3 @@
 # Be sure to restart your server when you modify this file.
 
-Openarchem::Application.config.session_store :cookie_store, key: '_openarchem_session'
+Rails.application.config.session_store :cookie_store, key: '_openarchem2_session'
